% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.are_weights}
\alias{.are_weights}
\title{Check if weights are valid (similar to datawizard)}
\usage{
.are_weights(weights)
}
\arguments{
\item{weights}{Numeric vector of weights}
}
\value{
Logical indicating if weights are valid
}
\description{
Check if weights are valid (similar to datawizard)
}
\keyword{internal}
