% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova.R
\name{.build_ancova_table}
\alias{.build_ancova_table}
\title{Build ANCOVA table from lm model and drop1 results}
\usage{
.build_ancova_table(
  model,
  type3,
  between_names,
  covariate_names,
  dv_name,
  n_total,
  w_name
)
}
\description{
Build ANCOVA table from lm model and drop1 results
}
\keyword{internal}
