% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{.calculate_ungrouped_stats}
\alias{.calculate_ungrouped_stats}
\title{Calculate statistics for ungrouped data}
\usage{
.calculate_ungrouped_stats(
  data,
  vars,
  weights_info,
  show,
  probs,
  na.rm,
  excess
)
}
\description{
Calculate statistics for ungrouped data
}
\keyword{internal}
