% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{.calculate_variable_stats}
\alias{.calculate_variable_stats}
\title{Calculate all requested statistics for a single variable}
\usage{
.calculate_variable_stats(x, w, var_name, show, probs, na.rm, excess)
}
\description{
Calculate all requested statistics for a single variable
}
\keyword{internal}
