% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{.classify_type}
\alias{.classify_type}
\title{Classify a variable's type for codebook display}
\usage{
.classify_type(x)
}
\arguments{
\item{x}{A data vector}
}
\value{
List with type (full description) and type_short (abbreviated)
}
\description{
Classify a variable's type for codebook display
}
\keyword{internal}
