% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{.codebook_to_html}
\alias{.codebook_to_html}
\title{Build HTML codebook from result object}
\usage{
.codebook_to_html(result)
}
\arguments{
\item{result}{A codebook result list}
}
\value{
An htmltools tag object
}
\description{
Build HTML codebook from result object
}
\keyword{internal}
