% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova.R
\name{.compute_estimated_marginal_means}
\alias{.compute_estimated_marginal_means}
\title{Compute estimated marginal means (adjusted for covariates at their grand mean)}
\usage{
.compute_estimated_marginal_means(
  model,
  data,
  dv_name,
  between_names,
  covariate_names,
  w_name
)
}
\description{
Compute estimated marginal means (adjusted for covariates at their grand mean)
}
\keyword{internal}
