% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_anova.R
\name{.compute_factorial_descriptives}
\alias{.compute_factorial_descriptives}
\title{Compute descriptive statistics for factorial design}
\usage{
.compute_factorial_descriptives(data, dv_name, between_names, w_name)
}
\description{
Compute descriptive statistics for factorial design
}
\keyword{internal}
