% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_anova.R
\name{.compute_factorial_levene}
\alias{.compute_factorial_levene}
\title{Compute Levene's test for factorial design}
\usage{
.compute_factorial_levene(data, dv_name, between_names, w_name)
}
\description{
Compute Levene's test for factorial design
}
\keyword{internal}
