% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa.R
\name{.compute_kmo}
\alias{.compute_kmo}
\title{Compute Kaiser-Meyer-Olkin Measure of Sampling Adequacy}
\usage{
.compute_kmo(cor_mat)
}
\description{
KMO measures the proportion of variance among variables that might be
common variance. SPSS-compatible implementation.
}
\keyword{internal}
