% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancova.R
\name{.compute_parameter_estimates}
\alias{.compute_parameter_estimates}
\title{Compute parameter estimates from the model}
\usage{
.compute_parameter_estimates(model, w_name)
}
\description{
Compute parameter estimates from the model
}
\keyword{internal}
