% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{.create_effect_size_df}
\alias{.create_effect_size_df}
\title{Create effect size data frame}
\usage{
.create_effect_size_df(
  var_name,
  cohens_d_val,
  hedges_g_val,
  glass_delta_val,
  is_paired = FALSE
)
}
\description{
Create effect size data frame
}
\keyword{internal}
