% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{.create_output_df}
\alias{.create_output_df}
\title{Create formatted output data frame for printing}
\usage{
.create_output_df(results_df, variables, show, is_weighted)
}
\description{
Create formatted output data frame for printing
}
\keyword{internal}
