% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa.R
\name{.efa_core}
\alias{.efa_core}
\title{Compute EFA for a single group}
\usage{
.efa_core(
  data,
  var_names,
  weights_vec,
  n_factors,
  rotation,
  extraction,
  use,
  sort,
  blank,
  na.rm
)
}
\description{
Compute EFA for a single group
}
\keyword{internal}
