% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa.R
\name{.efa_listwise_cor}
\alias{.efa_listwise_cor}
\title{Compute listwise correlation matrix}
\usage{
.efa_listwise_cor(data, var_names, weights_vec, na.rm)
}
\description{
Compute listwise correlation matrix
}
\keyword{internal}
