% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa.R
\name{.efa_pairwise_cor}
\alias{.efa_pairwise_cor}
\title{Compute pairwise correlation matrix (SPSS default for FACTOR)}
\usage{
.efa_pairwise_cor(data, var_names, weights_vec)
}
\description{
Compute pairwise correlation matrix (SPSS default for FACTOR)
}
\keyword{internal}
