% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R, R/helpers.R
\name{.effective_n}
\alias{.effective_n}
\title{Calculate effective sample size for weighted data}
\usage{
.effective_n(weights)

.effective_n(weights)
}
\arguments{
\item{weights}{Numeric vector of weights}
}
\value{
Effective sample size
}
\description{
Calculate effective sample size for weighted data

Calculate effective sample size for weighted data
}
\keyword{internal}
