% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{.extract_var_metadata}
\alias{.extract_var_metadata}
\title{Extract codebook metadata for a single variable}
\usage{
.extract_var_metadata(x, var_name, position, max_values = 10)
}
\arguments{
\item{x}{Data vector}

\item{var_name}{Variable name}

\item{position}{Column position in original data}

\item{max_values}{Maximum number of values to display before truncating}
}
\value{
Named list of metadata
}
\description{
Extract codebook metadata for a single variable
}
\keyword{internal}
