% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.get_border}
\alias{.get_border}
\title{Calculate dynamic border width based on data frame output}
\usage{
.get_border(df)
}
\arguments{
\item{df}{Data frame to be printed}
}
\value{
Character string with appropriate number of dashes
}
\description{
Calculate dynamic border width based on data frame output
}
\keyword{internal}
