% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{.get_t_test_border}
\alias{.get_t_test_border}
\title{Get table border for t-test formatting (similar to describe function)}
\usage{
.get_t_test_border(df)
}
\description{
Get table border for t-test formatting (similar to describe function)
}
\keyword{internal}
