% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{.glm_core}
\alias{.glm_core}
\title{Core logistic regression computation}
\usage{
.glm_core(data, formula, dep_name, pred_names, weights_vec, conf.level)
}
\description{
Core logistic regression computation
}
\keyword{internal}
