% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{.hosmer_lemeshow_test}
\alias{.hosmer_lemeshow_test}
\title{Hosmer-Lemeshow goodness-of-fit test}
\usage{
.hosmer_lemeshow_test(observed, predicted, weights_vec = NULL, n_groups = 10)
}
\description{
Hosmer-Lemeshow goodness-of-fit test
}
\keyword{internal}
