% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{.lm_anova}
\alias{.lm_anova}
\title{Compute ANOVA table for linear regression}
\usage{
.lm_anova(model, model_summary)
}
\description{
Compute ANOVA table for linear regression
}
\keyword{internal}
