% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{.lm_coefficients}
\alias{.lm_coefficients}
\title{Compute coefficients table with standardized coefficients}
\usage{
.lm_coefficients(
  model,
  model_summary,
  data,
  dep_name,
  pred_names,
  weights_vec,
  standardized,
  conf.level
)
}
\description{
Compute coefficients table with standardized coefficients
}
\keyword{internal}
