% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{.lm_core}
\alias{.lm_core}
\title{Core linear regression computation}
\usage{
.lm_core(
  data,
  formula,
  dep_name,
  pred_names,
  weights_vec,
  standardized,
  conf.level
)
}
\description{
Core linear regression computation
}
\keyword{internal}
