% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{.lm_core_pairwise}
\alias{.lm_core_pairwise}
\title{Core pairwise regression computation (matching SPSS /MISSING PAIRWISE)}
\usage{
.lm_core_pairwise(
  data,
  dep_name,
  pred_names,
  weights_vec,
  standardized,
  conf.level
)
}
\description{
Core pairwise regression computation (matching SPSS /MISSING PAIRWISE)
}
\keyword{internal}
