% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa.R
\name{.ml_max_factors}
\alias{.ml_max_factors}
\title{Compute maximum number of factors for ML extraction}
\usage{
.ml_max_factors(p)
}
\description{
ML requires non-negative degrees of freedom:
df = ((p - f)^2 - p - f) / 2 >= 0
}
\keyword{internal}
