% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_anova.R
\name{.parse_between_factors}
\alias{.parse_between_factors}
\title{Parse between-subjects factor specification}
\usage{
.parse_between_factors(between_quo, data)
}
\arguments{
\item{between_quo}{Quosure from enquo(between)}

\item{data}{Data frame for validation}
}
\value{
Character vector of factor names
}
\description{
Parse between-subjects factor specification
}
\keyword{internal}
