% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.print_border}
\alias{.print_border}
\title{Print dynamic border based on data frame width}
\usage{
.print_border(df)
}
\arguments{
\item{df}{Data frame that was just printed}
}
\description{
Print dynamic border based on data frame width
}
\keyword{internal}
