% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{.print_classification_table}
\alias{.print_classification_table}
\title{Print classification table}
\usage{
.print_classification_table(cls)
}
\description{
Print classification table
}
\keyword{internal}
