% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{.print_coefficients_table}
\alias{.print_coefficients_table}
\title{Print coefficients table}
\usage{
.print_coefficients_table(coefs, show_beta)
}
\description{
Print coefficients table
}
\keyword{internal}
