% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_cor.R
\name{.print_correlation_matrix}
\alias{.print_correlation_matrix}
\title{Internal function to print correlation matrices without cutoff}
\usage{
.print_correlation_matrix(
  mat,
  digits = 3,
  title = "Correlation Matrix:",
  type = "correlation"
)
}
\description{
Internal function to print correlation matrices without cutoff
}
\keyword{internal}
