% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{.print_descriptives_table}
\alias{.print_descriptives_table}
\title{Print descriptive statistics table}
\usage{
.print_descriptives_table(desc)
}
\description{
Print descriptive statistics table
}
\keyword{internal}
