% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunn_test.R
\name{.print_dunn_table}
\alias{.print_dunn_table}
\title{Print a Dunn test comparison table}
\usage{
.print_dunn_table(var_results, digits = 3)
}
\arguments{
\item{var_results}{Data frame with comparison results for one variable}

\item{digits}{Number of decimal places}
}
\description{
Print a Dunn test comparison table
}
\keyword{internal}
