% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa.R
\name{.print_efa_ungrouped}
\alias{.print_efa_ungrouped}
\title{Print EFA results for ungrouped data}
\usage{
.print_efa_ungrouped(x, digits = 3)
}
\description{
Print EFA results for ungrouped data
}
\keyword{internal}
