% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.print_group_header}
\alias{.print_group_header}
\title{Print group header}
\usage{
.print_group_header(group_info, group_vars)
}
\arguments{
\item{group_info}{Data frame with group information}

\item{group_vars}{Character vector of grouping variable names}
}
\description{
Print group header
}
\keyword{internal}
