% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{.print_grouped_results}
\alias{.print_grouped_results}
\title{Print results for grouped data}
\usage{
.print_grouped_results(x, is_weighted)
}
\description{
Print results for grouped data
}
\keyword{internal}
