% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.print_header}
\alias{.print_header}
\title{Print header for statistics}
\usage{
.print_header(statistic_name)
}
\arguments{
\item{statistic_name}{Name of the statistic (can include "Weighted" prefix)}
}
\description{
Print header for statistics
}
\keyword{internal}
