% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_tau.R
\name{.print_kendall_compact}
\alias{.print_kendall_compact}
\title{Print compact kendall_tau output for one group or ungrouped}
\usage{
.print_kendall_compact(x, corrs, weighted_tag, digits)
}
\description{
Print compact kendall_tau output for one group or ungrouped
}
\keyword{internal}
