% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_tau.R
\name{.print_kendall_verbose}
\alias{.print_kendall_verbose}
\title{Print verbose kendall_tau output for one group}
\usage{
.print_kendall_verbose(x, corrs, matrix_idx, show_cor, show_p, show_n, digits)
}
\description{
Print verbose kendall_tau output for one group
}
\keyword{internal}
