% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{.print_lm_ungrouped}
\alias{.print_lm_ungrouped}
\title{Print ungrouped linear regression result}
\usage{
.print_lm_ungrouped(x)
}
\description{
Print ungrouped linear regression result
}
\keyword{internal}
