% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efa.R
\name{.print_loading_matrix}
\alias{.print_loading_matrix}
\title{Print a loading matrix with blank suppression and optional sorting}
\usage{
.print_loading_matrix(mat, blank = 0.4, sort_loadings = TRUE, digits = 3)
}
\description{
Print a loading matrix with blank suppression and optional sorting
}
\keyword{internal}
