% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{.print_logistic_grouped}
\alias{.print_logistic_grouped}
\title{Print grouped logistic regression results}
\usage{
.print_logistic_grouped(x)
}
\description{
Print grouped logistic regression results
}
\keyword{internal}
