% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{.print_logistic_model_summary}
\alias{.print_logistic_model_summary}
\title{Print logistic model summary}
\usage{
.print_logistic_model_summary(ms)
}
\description{
Print logistic model summary
}
\keyword{internal}
