% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{.print_omnibus_test}
\alias{.print_omnibus_test}
\title{Print omnibus test of model coefficients}
\usage{
.print_omnibus_test(omnibus)
}
\description{
Print omnibus test of model coefficients
}
\keyword{internal}
