% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_cor.R
\name{.print_pearson_compact}
\alias{.print_pearson_compact}
\title{Print compact pearson_cor output for one group or ungrouped}
\usage{
.print_pearson_compact(x, corrs, weighted_tag, digits)
}
\description{
Print compact pearson_cor output for one group or ungrouped
}
\keyword{internal}
