% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{.print_reliability_grouped}
\alias{.print_reliability_grouped}
\title{Print reliability results for grouped data}
\usage{
.print_reliability_grouped(x, digits = 3)
}
\description{
Print reliability results for grouped data
}
\keyword{internal}
