% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{.print_reliability_ungrouped}
\alias{.print_reliability_ungrouped}
\title{Print reliability results for ungrouped data}
\usage{
.print_reliability_ungrouped(x, digits = 3)
}
\description{
Print reliability results for ungrouped data
}
\keyword{internal}
