% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearman_rho.R
\name{.print_spearman_compact}
\alias{.print_spearman_compact}
\title{Print compact spearman_rho output for one group or ungrouped}
\usage{
.print_spearman_compact(x, corrs, weighted_tag, digits)
}
\description{
Print compact spearman_rho output for one group or ungrouped
}
\keyword{internal}
