% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearman_rho.R
\name{.print_spearman_verbose}
\alias{.print_spearman_verbose}
\title{Print verbose spearman_rho output for one group}
\usage{
.print_spearman_verbose(x, corrs, matrix_idx, show_cor, show_p, show_n, digits)
}
\description{
Print verbose spearman_rho output for one group
}
\keyword{internal}
