% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{.print_summary_lm_grouped}
\alias{.print_summary_lm_grouped}
\title{Print grouped linear regression summary (verbose)}
\usage{
.print_summary_lm_grouped(x)
}
\description{
Print grouped linear regression summary (verbose)
}
\keyword{internal}
