% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{.print_summary_lm_ungrouped}
\alias{.print_summary_lm_ungrouped}
\title{Print ungrouped linear regression summary (verbose)}
\usage{
.print_summary_lm_ungrouped(x)
}
\description{
Print ungrouped linear regression summary (verbose)
}
\keyword{internal}
