% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{.print_summary_logistic_grouped}
\alias{.print_summary_logistic_grouped}
\title{Print grouped logistic regression summary (verbose)}
\usage{
.print_summary_logistic_grouped(x)
}
\description{
Print grouped logistic regression summary (verbose)
}
\keyword{internal}
