% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{.print_summary_logistic_ungrouped}
\alias{.print_summary_logistic_ungrouped}
\title{Print ungrouped logistic regression summary (verbose)}
\usage{
.print_summary_logistic_ungrouped(x)
}
\description{
Print ungrouped logistic regression summary (verbose)
}
\keyword{internal}
