% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_tau.R
\name{.print_tau_matrix}
\alias{.print_tau_matrix}
\title{Internal function to print correlation matrices without cutoff}
\usage{
.print_tau_matrix(
  mat,
  digits = 3,
  title = "Kendall's Tau Matrix:",
  type = "correlation"
)
}
\description{
Internal function to print correlation matrices without cutoff
}
\keyword{internal}
