% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{.print_ungrouped_results}
\alias{.print_ungrouped_results}
\title{Print results for ungrouped data}
\usage{
.print_ungrouped_results(x, is_weighted)
}
\description{
Print results for ungrouped data
}
\keyword{internal}
