% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.relabel_matrix}
\alias{.relabel_matrix}
\title{Relabel a matrix's dimnames using label maps}
\usage{
.relabel_matrix(mat, label_maps, max_width = 20)
}
\arguments{
\item{mat}{A matrix with dimnames}

\item{label_maps}{Named list of label maps (one per dimension)}

\item{max_width}{Maximum character width for truncation}
}
\value{
Matrix with relabelled dimnames
}
\description{
Relabel a matrix's dimnames using label maps
}
\keyword{internal}
