% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{.validate_and_clean_weights}
\alias{.validate_and_clean_weights}
\title{Validate and clean weights - central validation function}
\usage{
.validate_and_clean_weights(x, weights, na.rm = TRUE)
}
\description{
Validate and clean weights - central validation function
}
\keyword{internal}
