% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.validate_weights}
\alias{.validate_weights}
\title{Validate weights (similar to datawizard)}
\usage{
.validate_weights(weights, verbose = TRUE)
}
\arguments{
\item{weights}{Numeric vector of weights}

\item{verbose}{Logical; if TRUE, show warnings}
}
\value{
Logical indicating if weights are positive
}
\description{
Validate weights (similar to datawizard)
}
\keyword{internal}
