% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{.w_median}
\alias{.w_median}
\title{Weighted median using cumulative weights approach}
\usage{
.w_median(x, weights = NULL, na.rm = TRUE)
}
\description{
Weighted median using cumulative weights approach
}
\keyword{internal}
