% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{.w_mode}
\alias{.w_mode}
\title{Weighted mode (most frequent value by weight)}
\usage{
.w_mode(x, weights = NULL, na.rm = TRUE)
}
\description{
Weighted mode (most frequent value by weight)
}
\keyword{internal}
