% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{.weighted_cor}
\alias{.weighted_cor}
\title{Compute weighted correlation matrix from weighted covariance}
\usage{
.weighted_cor(mat, w)
}
\description{
Compute weighted correlation matrix from weighted covariance
}
\keyword{internal}
