% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reliability.R
\name{.weighted_cor_vec}
\alias{.weighted_cor_vec}
\title{Compute weighted correlation between two vectors}
\usage{
.weighted_cor_vec(x, y, w)
}
\description{
Compute weighted correlation between two vectors
}
\keyword{internal}
