% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_helpers.R
\name{format_number}
\alias{format_number}
\title{Format numeric value with appropriate decimal places}
\usage{
format_number(x, digits = 3, scientific = FALSE)
}
\arguments{
\item{x}{Numeric value}

\item{digits}{Number of decimal places}

\item{scientific}{Whether to use scientific notation for small p-values}
}
\description{
Format numeric value with appropriate decimal places
}
\keyword{internal}
