% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_helpers.R
\name{format_variable_name}
\alias{format_variable_name}
\title{Format variable name or label}
\usage{
format_variable_name(var, label = NULL)
}
\arguments{
\item{var}{Variable name}

\item{label}{Optional label}
}
\description{
Format variable name or label
}
\keyword{internal}
