% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_helpers.R
\name{get_standard_title}
\alias{get_standard_title}
\title{Standardize title based on weights and test type}
\usage{
get_standard_title(test_name, weights = NULL, suffix = "Results")
}
\arguments{
\item{test_name}{Name of the statistical test}

\item{weights}{Weights variable name or NULL}

\item{suffix}{"Results" or "Statistics"}
}
\description{
Standardize title based on weights and test type
}
\keyword{internal}
