% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_helpers.R
\name{get_table_width}
\alias{get_table_width}
\title{Calculate dynamic table width}
\usage{
get_table_width(df, min_width = 40)
}
\arguments{
\item{df}{Data frame to be printed}

\item{min_width}{Minimum table width}
}
\description{
Calculate dynamic table width
}
\keyword{internal}
