% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_value_labels}
\alias{get_value_labels}
\title{Get value labels for a variable}
\usage{
get_value_labels(x, freq_names)
}
\arguments{
\item{x}{The data vector (factor or haven-labelled)}

\item{freq_names}{Values to look up labels for}
}
\value{
Character vector of labels (or empty strings if no labels)
}
\description{
Get value labels for a variable
}
\keyword{internal}
