% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_square.R
\name{print.chi_square_results}
\alias{print.chi_square_results}
\title{Print method for chi_square_results}
\usage{
\method{print}{chi_square_results}(x, digits = 3, ...)
}
\arguments{
\item{x}{Chi-squared test results object}

\item{digits}{Number of decimal places (default: 3)}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for chi_square_results
}
