% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_tau.R
\name{print.kendall_tau_results}
\alias{print.kendall_tau_results}
\title{Print method for kendall_tau_results}
\usage{
\method{print}{kendall_tau_results}(x, digits = 3, ...)
}
\arguments{
\item{x}{A kendall_tau_results object}

\item{digits}{Number of decimal places to display (default: 3)}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for kendall_tau_results
}
