% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levene_test.R
\name{print.levene_test_results}
\alias{print.levene_test_results}
\title{Print method for Levene test results}
\usage{
\method{print}{levene_test_results}(x, digits = 3, ...)
}
\arguments{
\item{x}{A levene_test_results object}

\item{digits}{Number of decimal places to display}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for Levene test results
}
