% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spearman_rho.R
\name{print.spearman_rho_results}
\alias{print.spearman_rho_results}
\title{Print method for spearman_rho_results}
\usage{
\method{print}{spearman_rho_results}(x, digits = 3, ...)
}
\arguments{
\item{x}{An object of class "spearman_rho_results"}

\item{digits}{Number of decimal places to display}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for spearman_rho_results
}
