% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_quantile.R
\name{print.summary.w_quantile}
\alias{print.summary.w_quantile}
\title{Print summary of w_quantile results (with Kish's effective N)}
\usage{
\method{print}{summary.w_quantile}(x, ...)
}
\arguments{
\item{x}{A \code{summary.w_quantile} object.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print summary of w_quantile results (with Kish's effective N)
}
