% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_factory.R
\name{print.summary.w_statistic}
\alias{print.summary.w_statistic}
\title{Print summary of a w_* statistic (detailed output)}
\usage{
\method{print}{summary.w_statistic}(x, ...)
}
\arguments{
\item{x}{A \code{summary.w_statistic} object created by
\code{summary()} on a \code{w_*} result.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print summary of a w_* statistic (detailed output)
}
