% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{print.t_test_result}
\alias{print.t_test_result}
\title{Print method for t_test_result}
\usage{
\method{print}{t_test_result}(x, digits = 3, ...)
}
\arguments{
\item{x}{A t_test_result object}

\item{digits}{Number of decimal places to display}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for t_test_result
}
