% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_iqr.R
\name{print.w_iqr}
\alias{print.w_iqr}
\title{Print method for w_iqr objects}
\usage{
\method{print}{w_iqr}(x, digits = 3, ...)
}
\arguments{
\item{x}{A w_iqr object}

\item{digits}{Number of decimal places to display}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for w_iqr objects
}
