% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_mean.R
\name{print.w_mean}
\alias{print.w_mean}
\title{Print method for w_mean objects}
\usage{
\method{print}{w_mean}(x, digits = 3, ...)
}
\arguments{
\item{x}{A w_mean object}

\item{digits}{Number of decimal places to display}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for w_mean objects
}
