% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_median.R
\name{print.w_median}
\alias{print.w_median}
\title{Print method for w_median objects}
\usage{
\method{print}{w_median}(x, digits = 3, ...)
}
\arguments{
\item{x}{A w_median object}

\item{digits}{Number of decimal places to display}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for w_median objects
}
