% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_modus.R
\name{print.w_modus}
\alias{print.w_modus}
\title{Print method for w_modus objects}
\usage{
\method{print}{w_modus}(x, digits = 3, ...)
}
\arguments{
\item{x}{A w_modus object}

\item{digits}{Number of decimal places to display}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for w_modus objects
}
