% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_range.R
\name{print.w_range}
\alias{print.w_range}
\title{Print method for w_range objects}
\usage{
\method{print}{w_range}(x, digits = 3, ...)
}
\arguments{
\item{x}{A w_range object}

\item{digits}{Number of decimal places to display}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for w_range objects
}
