% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_sd.R
\name{print.w_sd}
\alias{print.w_sd}
\title{Print method for w_sd objects}
\usage{
\method{print}{w_sd}(x, digits = 3, ...)
}
\arguments{
\item{x}{A w_sd object}

\item{digits}{Number of decimal places to display}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for w_sd objects
}
