% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_se.R
\name{print.w_se}
\alias{print.w_se}
\title{Print method for w_se objects}
\usage{
\method{print}{w_se}(x, digits = 3, ...)
}
\arguments{
\item{x}{A w_se object}

\item{digits}{Number of decimal places to display}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for w_se objects
}
