% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_skew.R
\name{print.w_skew}
\alias{print.w_skew}
\title{Print method for w_skew objects}
\usage{
\method{print}{w_skew}(x, digits = 3, ...)
}
\arguments{
\item{x}{A w_skew object}

\item{digits}{Number of decimal places to display}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for w_skew objects
}
