% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w_var.R
\name{print.w_var}
\alias{print.w_var}
\title{Print method for w_var objects}
\usage{
\method{print}{w_var}(x, digits = 3, ...)
}
\arguments{
\item{x}{A w_var object}

\item{digits}{Number of decimal places to display}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for w_var objects
}
