% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_helpers.R
\name{print_group_header}
\alias{print_group_header}
\title{Print grouped data header}
\usage{
print_group_header(group_values, prefix = "Group")
}
\arguments{
\item{group_values}{Named vector or data frame of group values}

\item{prefix}{Text to print before group info}
}
\description{
Print grouped data header
}
\keyword{internal}
