% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_helpers.R
\name{print_info_section}
\alias{print_info_section}
\title{Print test information section}
\usage{
print_info_section(info, indent = 0)
}
\arguments{
\item{info}{Named list of information to display}

\item{indent}{Number of spaces to indent}
}
\description{
Print test information section
}
\keyword{internal}
