% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_helpers.R
\name{print_significance_legend}
\alias{print_significance_legend}
\title{Print significance codes legend}
\usage{
print_significance_legend(show = TRUE)
}
\arguments{
\item{show}{Logical, whether to show the legend}
}
\description{
Print significance codes legend
}
\keyword{internal}
