% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_helpers.R
\name{print_test_parameters}
\alias{print_test_parameters}
\title{Print standard test parameters}
\usage{
print_test_parameters(params)
}
\arguments{
\item{params}{List of parameters (conf.level, alternative, etc.)}
}
\description{
Print standard test parameters
}
\keyword{internal}
