% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{.lm_descriptives}
\alias{.lm_descriptives}
\title{Compute descriptive statistics for regression variables}
\usage{
.lm_descriptives(data, var_names, weights_vec)
}
\description{
Compute descriptive statistics for regression variables
}
\keyword{internal}
