% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{.w_kurtosis}
\alias{.w_kurtosis}
\title{Weighted kurtosis with bias correction}
\usage{
.w_kurtosis(x, weights = NULL, na.rm = TRUE, excess = TRUE)
}
\description{
Weighted kurtosis with bias correction
}
\keyword{internal}
