% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{.w_quantile}
\alias{.w_quantile}
\title{Weighted quantiles using cumulative weights}
\usage{
.w_quantile(x, weights = NULL, probs = 0.5, na.rm = TRUE)
}
\description{
Weighted quantiles using cumulative weights
}
\keyword{internal}
