% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mariposa-package.R
\docType{package}
\name{mariposa-package}
\alias{mariposa}
\alias{mariposa-package}
\title{mariposa: Marburg Initiative for Political and Social Analysis - Statistical Tools for Survey Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Professional statistical analysis of survey data with full support for survey weights, grouped operations, and tidyverse integration. Provides 27 functions for descriptive statistics, hypothesis testing, correlation analysis, post-hoc comparisons, and weighted statistics. All results are validated against SPSS v29 for reproducibility. Designed for survey researchers, social scientists, and students working with complex survey designs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YannickDiehl/mariposa}
  \item Report bugs at \url{https://github.com/YannickDiehl/mariposa/issues}
}

}
\author{
\strong{Maintainer}: Yannick Diehl \email{yannick.diehl@gmail.com}

}
\keyword{internal}
