% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mariposa-package.R
\docType{package}
\name{mariposa-package}
\alias{mariposa}
\alias{mariposa-package}
\title{mariposa: Marburg Initiative for Political and Social Analysis - Statistical Tools for Survey Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A comprehensive package for statistical analysis of survey data, providing tools for data cleaning, weighting, complex survey design analysis, and visualization. Designed to handle common survey methodologies and statistical procedures used in survey research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YannickDiehl/mariposa}
  \item Report bugs at \url{https://github.com/YannickDiehl/mariposa/issues}
}

}
\author{
\strong{Maintainer}: Yannick Diehl \email{yannick.diehl@gmail.com}

}
\keyword{internal}
