% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spss.R
\name{na_frequencies}
\alias{na_frequencies}
\title{Frequency Table of Missing Value Types}
\usage{
na_frequencies(x)
}
\arguments{
\item{x}{A numeric vector with tagged NAs (from \code{\link[=read_spss]{read_spss()}}).}
}
\value{
A data frame with columns:
\item{tag}{The tag character (a-z, A-Z, 0-9)}
\item{n}{Number of cases with this missing type}
\item{spss_code}{The original SPSS missing value code}
\item{label}{The SPSS value label for this missing type}
}
\description{
Shows a breakdown of the different types of missing values in a variable
that was read with \code{\link[=read_spss]{read_spss()}} using \code{tag.na = TRUE}.
}
\examples{
\dontrun{
data <- read_spss("survey.sav")
na_frequencies(data$satisfaction)
#   tag    n spss_code            label
# 1   b 1774       -11       TNZ: SPLIT
# 2   c   63        -9     KEINE ANGABE
# 3   d   11        -8 WEISS NICHT
# 4   a    6       -42 DATENFEHLER: MFN
}

}
\seealso{
\code{\link[=read_spss]{read_spss()}}, \code{\link[=untag_na]{untag_na()}}, \code{\link[=strip_tags]{strip_tags()}}

Other spss: 
\code{\link{read_spss}()},
\code{\link{strip_tags}()},
\code{\link{untag_na}()}
}
\concept{spss}
