% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{print.crosstab}
\alias{print.crosstab}
\title{Print method for crosstab results}
\usage{
\method{print}{crosstab}(x, ...)
}
\arguments{
\item{x}{A crosstab result object}

\item{...}{Additional arguments (currently unused)}
}
\description{
Print method for crosstab results
}
