% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kruskal_wallis.R
\name{print.kruskal_wallis}
\alias{print.kruskal_wallis}
\title{Print method for Kruskal-Wallis test results}
\usage{
\method{print}{kruskal_wallis}(x, digits = 3, ...)
}
\arguments{
\item{x}{A kruskal_wallis object}

\item{digits}{Number of decimal places to display (default: 3)}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for Kruskal-Wallis test results
}
