% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{print.linear_regression}
\alias{print.linear_regression}
\title{Print Linear Regression Results}
\usage{
\method{print}{linear_regression}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"linear_regression"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Linear Regression Results
}
