% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{print.logistic_regression}
\alias{print.logistic_regression}
\title{Print Logistic Regression Results}
\usage{
\method{print}{logistic_regression}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"logistic_regression"}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print Logistic Regression Results
}
