% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mann_whitney.R
\name{print.mann_whitney}
\alias{print.mann_whitney}
\title{Print method for Mann-Whitney test results}
\usage{
\method{print}{mann_whitney}(x, digits = 3, ...)
}
\arguments{
\item{x}{A mann_whitney object}

\item{digits}{Number of decimal places to display}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for Mann-Whitney test results
}
