% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcnemar_test.R
\name{print.mcnemar_test}
\alias{print.mcnemar_test}
\title{Print McNemar test results}
\usage{
\method{print}{mcnemar_test}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"mcnemar_test"}}

\item{digits}{Number of decimal places (default: 4)}

\item{...}{Additional arguments (currently unused)}
}
\value{
Invisibly returns the input object \code{x}.
}
\description{
Print McNemar test results
}
