% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearson_cor.R
\name{print.pearson_cor}
\alias{print.pearson_cor}
\title{Print method for pearson_cor}
\usage{
\method{print}{pearson_cor}(x, digits = 3, ...)
}
\arguments{
\item{x}{A pearson_cor object}

\item{digits}{Number of decimal places to display (default: 3)}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for pearson_cor
}
