% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcoxon_test.R
\name{print.wilcoxon_test}
\alias{print.wilcoxon_test}
\title{Print method for Wilcoxon signed-rank test results}
\usage{
\method{print}{wilcoxon_test}(x, digits = 3, ...)
}
\arguments{
\item{x}{A wilcoxon_test object}

\item{digits}{Number of decimal places to display (default: 3)}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for Wilcoxon signed-rank test results
}
