% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_helpers.R
\name{print_separator}
\alias{print_separator}
\title{Print horizontal separator line}
\usage{
print_separator(width = 40, char = "-")
}
\arguments{
\item{width}{Width of the line}

\item{char}{Character to use for the line}
}
\description{
Print horizontal separator line
}
\keyword{internal}
